includes/requirements.txt
pyyaml==5.3
prometheus_client==0.7.1
holidays==0.10.3
